refactor: restructure Header component into separate files and implement LanguageSwitcher for improved modularity
