+++ src/styles/colors.ts
@@ -6,6 +6,7 @@ export const colors = {
   // 브랜드 색상
   black: '#000000',
   white: '#FFFFFF',
+  headerBorder: '#F5F5F5',
 } as const;
 
 export type ColorType = keyof typeof colors;
